fix(stdlib): fix some of the memory leaks within the standard library
Fix some of the remaining leaks within the standard library and enable
tests to ensure that memory leaks are checked for every end to end test.

There were two functions that I could not figure out how to fix for
memory leaks. The `join` function was too complex for me to feel
comfortable with deciding when to release memory. The `cov` function
uses the `join` function so I similarly could not fix that one. These
tests are specially skipped until we are able to address the memory
leaks in `join` with a possible refactor.